Fix labels not positioning correctly after multiple block resizes (#13)
When changing width or height of a block, the adjustment of the text position was incorrectly using relative offsets vs the current position, rather than offsets relative to the starting position. This meant that the first size change in each axis worked correctly, but subsequent size changes would result in labels moving to an incorrect position.

Also add a small rotation to all block labels when created to force browsers not to pixel-snap them, which makes text animation very jittery.