Fix Scene.fastforward()
Scenes default to adding new animations to the end of timeline by
querying the current end time of the animation. In reality what we
need is the end time from the point of view of the animation
timeline (which GSAP calls the duration) not the actual end time
(after seeks and time scaling playback speeds). This fix flips
all use of tl.endTime() to tl.duration() to correct for that.